COM-2054 - Add helper for notification
