+++ src/AppContainer/index.tsx
@@ -1,9 +1,8 @@
 import React, { useEffect, useContext, useRef } from 'react';
-import { StatusBar, View, Platform } from 'react-native';
+import { StatusBar, View } from 'react-native';
 import { connect } from 'react-redux';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
-import Constants from 'expo-constants';
 import * as Notifications from 'expo-notifications';
 import Analytics from '../core/helpers/analytics';
 import asyncStorage from '../core/helpers/asyncStorage';
@@ -30,7 +29,11 @@ import styles from './styles';
 import PasswordEdition from '../screens/profile/PasswordEdition';
 import PasswordReset from '../screens/PasswordReset';
 import Home from '../Home';
-import { BLENDED_COURSE_INSCRIPTION } from '../core/data/constants';
+import {
+  registerForPushNotificationsAsync,
+  _handleNotification,
+  _handleNotificationResponse,
+} from '../core/helpers/notifications';
 
 const MainStack = createStackNavigator();
 
@@ -44,57 +47,6 @@ const AppContainer = ({ setLoggedUser, statusBarVisible }: AppContainerProps) =>
   const routeNameRef = useRef<string>();
   const notificationListener = useRef();
   const responseListener = useRef();
-
-  const registerForPushNotificationsAsync = async () => {
-    if (!Constants.isDevice) return console.log('Les notifications ne sont pas disponibles sur simulateur');
-
-    const { status: existingStatus } = await Notifications.getPermissionsAsync();
-    let finalStatus = existingStatus;
-    console.log('ici', existingStatus);
-
-    if (existingStatus !== 'granted') {
-      const { status } = await Notifications.requestPermissionsAsync();
-      finalStatus = status;
-      console.log('là', finalStatus);
-    }
-
-    if (finalStatus !== 'granted') {
-      console.log('Failed to get push token for push notification!');
-      return;
-    }
-
-    if (Platform.OS === 'android') {
-      Notifications.setNotificationChannelAsync('default', {
-        name: 'default',
-        importance: Notifications.AndroidImportance.MAX,
-        vibrationPattern: [0, 250, 250, 250],
-        lightColor: '#FF231F7C',
-      });
-    }
-
-    const token = (await Notifications.getExpoPushTokenAsync()).data;
-
-    return token;
-  };
-
-  const _handleNotification = (notification) => {
-    console.log('j\'ai reçu ceci : \n', notification);
-  };
-
-  const _handleNotificationResponse = (response) => {
-    const { type, _id } = response.notification.request.content.data;
-    console.log('type \n', type);
-    console.log('_id \n', _id);
-
-    switch (type) {
-      case BLENDED_COURSE_INSCRIPTION:
-        console.log('ici');
-        navigationRef.current?.navigate('CourseProfile', { courseId: _id });
-        break;
-      default:
-        return null;
-    }
-  };
 
   useEffect(() => {
     registerForPushNotificationsAsync().then(token => console.log('expoToken', token));
